+++ CHANGES.rst
@@ -6,6 +6,9 @@ Changes
 --------
 
 - More readable chat logs with topic change messages
+- Command ``/ts`` displays topics list and unread messages count
+- Command ``/t`` selects a topic *and* lists recent messages
+- Removed command ``/ls``
 - Recent topics are updated dynamically based on incoming messages
 - Input and output characters don't interfere anymore
 - Timestamps in the local timezone